+++ lib/src/models/models.dart
@@ -7,6 +7,8 @@ export 'desk_sidebar_items_response/desk_page.dart';
 export 'desk_sidebar_items_response/desk_sidebar_items_response.dart';
 export 'desktop_page_request.dart';
 export 'desktop_page_response/desktop_page_response.dart';
+export 'filter.dart';
+export 'frappe_doc.dart';
 export 'frappe_exception.dart';
 export 'get_count_request.dart';
 export 'get_count_response.dart';